requirements.txt
streamlit
pillow
mediapipe
opencv-python
numpy